+++ lib/kitchen/element_base.rb
@@ -754,7 +754,7 @@ module Kitchen
     def_delegators :as_enumerator, :pages, :chapters, :terms, :figures, :notes, :tables, :examples,
                    :metadatas, :non_introduction_pages, :units, :titles, :exercises, :references,
                    :composite_pages, :composite_chapters, :solutions, :injected_questions,
-                   :search_with
+                   :search_with, :sections
 
     # Returns this element as an enumerator (over only one element, itself)
     #